Taking a Different Approach
It's been a long several months! Now that I have time to sit back down and work on this project, here's what's going on.

I'm creating a separate project called "ProtocolSharp" which is a server-side assistant to utilizing the Minecraft protocol. Maybe a client-side version will come?

Anyways, while taking on this approach, one of the big stumbling blocks in my way is ENTITIES. And because I want Redstone to be powerful and mod/plugin worthy, I'm adding each entity and their metadatas in their own classes following the pattern on wiki.vg/Protocol. It's not hard to understand, but VERY tedious.