Updates to Entity.cs
This is the first part of updating Entities to work with NBT. You will notice added was some entity values that will be needed later. Also you will notice the field for the NBT Compound. It is not finished yet, but you can see the progress being made on it.