Simplified Blocks
After a bit of getting frustrated with the blocks and how things work, I decided to simplify and go with an easy solution. After a while of researching, I stumbled across a json file that contains extracted data from 1.16.4 on blocks. From here I now use the System.Text.Json class to read the blocks on call.