Migrate to using new/updated libraries
Redstone will use SmartBlocks as it's world processing, but in order to do that will need to have access to the same integral types (VarInt, VarLong, Angle, Position, etc.). This meant a need to create a library for types like these was necessary. This also prompted me to update SmartNbt to 1.1, which in turn I've updated Redstone with as well.